requirements.txt
fastapi
uvicorn
slack-sdk
python-dotenv
requests
validators
pydantic_settings
python-multipart
aiohttp
pydantic
sentry_sdk
